Upgrade

+++ MaviSoftServerV1.0/PanelLog.cs
@@ -1044,7 +1044,7 @@ namespace MaviSoftServerV1._0
                                 {
                                     if (TLPR.Trim() == "")
                                     {
-                                        if (int.Parse(TCardID) == 0 || TCardID.Trim() == "")
+                                        if (long.Parse(TCardID) == 0 || TCardID.Trim() == "")
                                             break;
                                     }
                                 }
